DFR-2624-update test variables
